Create Movie Collection App
With this code, a user will be able to add new movies to their collection to keep track of them and list all the movies in their collection to see what is already there. A user will also be able to search for a movie by its title easily for when the collection grows.